upport FileReaders in massage_in and massage_out.
Since it's pretty common that a file will contain chunks in several different
formats, in which case each may be processed with its own FileReader, I've
added a little code that allows the user to assign a FileReader directly in
the massage_in and massage_out dictionaries, rather than fr.pack and fr.unpack.
This should, I hope, make it clear when that is happening.

This should not change the behavior of FileReaders as previously used.